--- project_app/src/_services/mockSimulations.tsx
@@ -1,1093 +1,0 @@
-"use client";
-import { ColumnDef } from "@tanstack/react-table";
-
-export type SimulatioData = {
-  id: string;
-  nome: string;
-  email: string;
-  telefone: string;
-  codigoDaUnidadeConsumidora: string;
-  enquadramento: string;
-  modeloFasico: string;
-  valor: number;
-};
-
-export const columns: ColumnDef<SimulatioData>[] = [
-  {
-    accessorKey: "nome",
-    header: "Nome",
-    size: 160,
-    cell: ({ row }) => (
-      <div className="truncate w-[160px] text-center">
-        {row.getValue("nome")}
-      </div>
-    ),
-  },
-  {
-    accessorKey: "email",
-    header: "Email",
-    size: 200,
-    cell: ({ row }) => (
-      <div className="truncate text-center w-[200px]">
-        {row.getValue("email")}
-      </div>
-    ),
-  },
-  {
-    accessorKey: "telefone",
-    header: "Telefone",
-    size: 130,
-    cell: ({ row }) => (
-      <div className="w-[130px] text-center">{row.getValue("telefone")}</div>
-    ),
-  },
-  {
-    accessorKey: "codigoDaUnidadeConsumidora",
-    header: "Cód. Unidade",
-    size: 160,
-    cell: ({ row }) => (
-      <div className="w-[160px] text-center">
-        {row.getValue("codigoDaUnidadeConsumidora")}
-      </div>
-    ),
-  },
-  {
-    accessorKey: "enquadramento",
-    header: "Enquadramento",
-    size: 120,
-    cell: ({ row }) => (
-      <div className="w-[120px] text-center">
-        {row.getValue("enquadramento")}
-      </div>
-    ),
-  },
-  {
-    accessorKey: "modeloFasico",
-    header: "Mod. Fásico",
-    size: 120,
-    cell: ({ row }) => (
-      <div className="w-[120px] text-center">
-        {row.getValue("modeloFasico")}
-      </div>
-    ),
-  },
-  {
-    accessorKey: "valor",
-    header: () => <div className="text-center w-[120px]">Valor</div>,
-    size: 100,
-    cell: ({ row }) => {
-      const valor = parseFloat(row.getValue("valor"));
-      const formatted = new Intl.NumberFormat("pt-BR", {
-        style: "currency",
-        currency: "BRL",
-      }).format(valor);
-
-      return (
-        <div className="text-right font-medium w-[100px]">{formatted}</div>
-      );
-    },
-  },
-];
-
-export const simulations = [
-  {
-    id: 1,
-    nome: "Kale Simenon",
-    email: "ksimenon0@nyu.edu",
-    telefone: "6895910407",
-    codigoDaUnidadeConsumidora: "7417232302",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 5377.54,
-  },
-  {
-    id: 2,
-    nome: "Calli Spire",
-    email: "cspire1@discuz.net",
-    telefone: "3985951115",
-    codigoDaUnidadeConsumidora: "9584909401",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 5748.39,
-  },
-  {
-    id: 3,
-    nome: "Madge Nussey",
-    email: "mnussey2@nbcnews.com",
-    telefone: "2999292470",
-    codigoDaUnidadeConsumidora: "5962994183",
-    enquadramento: "B1",
-    modeloFasico: "trifasico",
-    valor: 4493.62,
-  },
-  {
-    id: 4,
-    nome: "Madelina Starbeck",
-    email: "mstarbeck3@sciencedirect.com",
-    telefone: "2681630256",
-    codigoDaUnidadeConsumidora: "1776720318",
-    enquadramento: "B1",
-    modeloFasico: "monofasico",
-    valor: 5251.36,
-  },
-  {
-    id: 5,
-    nome: "Jill Lain",
-    email: "jlain4@webmd.com",
-    telefone: "3101658082",
-    codigoDaUnidadeConsumidora: "8163345845",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 3361.62,
-  },
-  {
-    id: 6,
-    nome: "Madeleine Couve",
-    email: "mcouve5@yale.edu",
-    telefone: "7226228324",
-    codigoDaUnidadeConsumidora: "5382590850",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 5547.79,
-  },
-  {
-    id: 7,
-    nome: "Cindee Brain",
-    email: "cbrain6@washington.edu",
-    telefone: "5012668514",
-    codigoDaUnidadeConsumidora: "0319481786",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 8174.1,
-  },
-  {
-    id: 8,
-    nome: "Sayre Witson",
-    email: "switson7@cam.ac.uk",
-    telefone: "7622209722",
-    codigoDaUnidadeConsumidora: "1204738386",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 6346.8,
-  },
-  {
-    id: 9,
-    nome: "Irma Woodfield",
-    email: "iwoodfield8@woothemes.com",
-    telefone: "6126560575",
-    codigoDaUnidadeConsumidora: "7534714214",
-    enquadramento: "AX",
-    modeloFasico: "monofasico",
-    valor: 5730.99,
-  },
-  {
-    id: 10,
-    nome: "Tod Chater",
-    email: "tchater9@360.cn",
-    telefone: "9011165601",
-    codigoDaUnidadeConsumidora: "1985411245",
-    enquadramento: "B1",
-    modeloFasico: "trifasico",
-    valor: 3054.48,
-  },
-  {
-    id: 11,
-    nome: "Oralie Stanner",
-    email: "ostannera@drupal.org",
-    telefone: "8932817502",
-    codigoDaUnidadeConsumidora: "6097824289",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 8639.55,
-  },
-  {
-    id: 12,
-    nome: "Sax Blackstone",
-    email: "sblackstoneb@cmu.edu",
-    telefone: "9154169212",
-    codigoDaUnidadeConsumidora: "1582084637",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 6810.51,
-  },
-  {
-    id: 13,
-    nome: "Christyna Kaindl",
-    email: "ckaindlc@topsy.com",
-    telefone: "3601244485",
-    codigoDaUnidadeConsumidora: "4276861969",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 5128.86,
-  },
-  {
-    id: 14,
-    nome: "Cynthie Cutten",
-    email: "ccuttend@friendfeed.com",
-    telefone: "9886662224",
-    codigoDaUnidadeConsumidora: "7480083521",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 4259.98,
-  },
-  {
-    id: 15,
-    nome: "Ruthie Swadlin",
-    email: "rswadline@drupal.org",
-    telefone: "4212424856",
-    codigoDaUnidadeConsumidora: "3767382539",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 3337.75,
-  },
-  {
-    id: 16,
-    nome: "Klaus Battrick",
-    email: "kbattrickf@yellowpages.com",
-    telefone: "4524712309",
-    codigoDaUnidadeConsumidora: "0930927680",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 4317.31,
-  },
-  {
-    id: 17,
-    nome: "Aldin Moncrefe",
-    email: "amoncrefeg@buzzfeed.com",
-    telefone: "6717873391",
-    codigoDaUnidadeConsumidora: "4071512644",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 5679.82,
-  },
-  {
-    id: 18,
-    nome: "Rickie Rosenwasser",
-    email: "rrosenwasserh@photobucket.com",
-    telefone: "7055213992",
-    codigoDaUnidadeConsumidora: "6843167462",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 9761.5,
-  },
-  {
-    id: 19,
-    nome: "Datha Douse",
-    email: "ddousei@msu.edu",
-    telefone: "1162213456",
-    codigoDaUnidadeConsumidora: "3919225767",
-    enquadramento: "AX",
-    modeloFasico: "monofasico",
-    valor: 9070.0,
-  },
-  {
-    id: 20,
-    nome: "Collie Parkman",
-    email: "cparkmanj@blogspot.com",
-    telefone: "5945133837",
-    codigoDaUnidadeConsumidora: "5022479478",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 9318.37,
-  },
-  {
-    id: 21,
-    nome: "Diana Aneley",
-    email: "daneleyk@lycos.com",
-    telefone: "5699151786",
-    codigoDaUnidadeConsumidora: "5294878194",
-    enquadramento: "B1",
-    modeloFasico: "trifasico",
-    valor: 7061.87,
-  },
-  {
-    id: 22,
-    nome: "Gretta Toke",
-    email: "gtokel@wsj.com",
-    telefone: "8855056964",
-    codigoDaUnidadeConsumidora: "8215781284",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 3322.94,
-  },
-  {
-    id: 23,
-    nome: "Stevena Ashwell",
-    email: "sashwellm@reddit.com",
-    telefone: "7881618398",
-    codigoDaUnidadeConsumidora: "7418635788",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 6496.54,
-  },
-  {
-    id: 24,
-    nome: "Pietra Bleas",
-    email: "pbleasn@spiegel.de",
-    telefone: "4568051330",
-    codigoDaUnidadeConsumidora: "7075379496",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 8425.2,
-  },
-  {
-    id: 25,
-    nome: "Nickolai Sunner",
-    email: "nsunnero@yahoo.com",
-    telefone: "7383417290",
-    codigoDaUnidadeConsumidora: "0692326960",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 7569.54,
-  },
-  {
-    id: 26,
-    nome: "Kakalina Packer",
-    email: "kpackerp@salon.com",
-    telefone: "8653987502",
-    codigoDaUnidadeConsumidora: "4659316475",
-    enquadramento: "AX",
-    modeloFasico: "bifasico",
-    valor: 8578.43,
-  },
-  {
-    id: 27,
-    nome: "Garnet Galloway",
-    email: "ggallowayq@yahoo.com",
-    telefone: "2619593007",
-    codigoDaUnidadeConsumidora: "6130849915",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 3853.79,
-  },
-  {
-    id: 28,
-    nome: "Kalle Thecham",
-    email: "kthechamr@kickstarter.com",
-    telefone: "5576776867",
-    codigoDaUnidadeConsumidora: "1581017693",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 8054.84,
-  },
-  {
-    id: 29,
-    nome: "Heather Bellenie",
-    email: "hbellenies@japanpost.jp",
-    telefone: "1824337251",
-    codigoDaUnidadeConsumidora: "0460061313",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 8650.32,
-  },
-  {
-    id: 30,
-    nome: "Nerty Grimster",
-    email: "ngrimstert@usgs.gov",
-    telefone: "2323956568",
-    codigoDaUnidadeConsumidora: "5338682303",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 8592.52,
-  },
-  {
-    id: 31,
-    nome: "Donna Bebb",
-    email: "dbebbu@dailymotion.com",
-    telefone: "7252453262",
-    codigoDaUnidadeConsumidora: "4507460034",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 6469.91,
-  },
-  {
-    id: 32,
-    nome: "Mohandis Giannassi",
-    email: "mgiannassiv@nhs.uk",
-    telefone: "5303772662",
-    codigoDaUnidadeConsumidora: "3133978411",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 4912.67,
-  },
-  {
-    id: 33,
-    nome: "Tully Gallemore",
-    email: "tgallemorew@statcounter.com",
-    telefone: "2271938666",
-    codigoDaUnidadeConsumidora: "9784578468",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 9557.99,
-  },
-  {
-    id: 34,
-    nome: "Teri Ramage",
-    email: "tramagex@google.es",
-    telefone: "8004444616",
-    codigoDaUnidadeConsumidora: "5161877553",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 8442.73,
-  },
-  {
-    id: 35,
-    nome: "Kean Pierrepont",
-    email: "kpierreponty@drupal.org",
-    telefone: "1369098002",
-    codigoDaUnidadeConsumidora: "8582607520",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 5286.88,
-  },
-  {
-    id: 36,
-    nome: "Elsi Peach",
-    email: "epeachz@hibu.com",
-    telefone: "1646791658",
-    codigoDaUnidadeConsumidora: "3843763348",
-    enquadramento: "B1",
-    modeloFasico: "trifasico",
-    valor: 3830.69,
-  },
-  {
-    id: 37,
-    nome: "Elmira Springett",
-    email: "espringett10@xrea.com",
-    telefone: "1296520335",
-    codigoDaUnidadeConsumidora: "2937235115",
-    enquadramento: "AX",
-    modeloFasico: "bifasico",
-    valor: 7966.9,
-  },
-  {
-    id: 38,
-    nome: "Ezekiel Welfair",
-    email: "ewelfair11@opera.com",
-    telefone: "1003194665",
-    codigoDaUnidadeConsumidora: "6279139239",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 4127.25,
-  },
-  {
-    id: 39,
-    nome: "Taffy Kynder",
-    email: "tkynder12@netvibes.com",
-    telefone: "4608451292",
-    codigoDaUnidadeConsumidora: "8310394373",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 6734.97,
-  },
-  {
-    id: 40,
-    nome: "Witty Stable",
-    email: "wstable13@canalblog.com",
-    telefone: "2006725478",
-    codigoDaUnidadeConsumidora: "3590497114",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 9581.97,
-  },
-  {
-    id: 41,
-    nome: "Sonya Robertsen",
-    email: "srobertsen14@techcrunch.com",
-    telefone: "4382032317",
-    codigoDaUnidadeConsumidora: "1765771013",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 7953.42,
-  },
-  {
-    id: 42,
-    nome: "Karia Attryde",
-    email: "kattryde15@mapy.cz",
-    telefone: "5719564809",
-    codigoDaUnidadeConsumidora: "2990089365",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 8103.04,
-  },
-  {
-    id: 43,
-    nome: "Lynsey Lawn",
-    email: "llawn16@discuz.net",
-    telefone: "4434368534",
-    codigoDaUnidadeConsumidora: "8037315843",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 5659.1,
-  },
-  {
-    id: 44,
-    nome: "Rhianon Mechi",
-    email: "rmechi17@tiny.cc",
-    telefone: "6585729713",
-    codigoDaUnidadeConsumidora: "5245488442",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 4142.69,
-  },
-  {
-    id: 45,
-    nome: "Page Minichillo",
-    email: "pminichillo18@1688.com",
-    telefone: "1439663742",
-    codigoDaUnidadeConsumidora: "2347376101",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 7310.32,
-  },
-  {
-    id: 46,
-    nome: "Hamilton Lonergan",
-    email: "hlonergan19@google.ru",
-    telefone: "1509405169",
-    codigoDaUnidadeConsumidora: "8429297405",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 6506.14,
-  },
-  {
-    id: 47,
-    nome: "Niko Doll",
-    email: "ndoll1a@yellowbook.com",
-    telefone: "1771646383",
-    codigoDaUnidadeConsumidora: "0302271589",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 3113.57,
-  },
-  {
-    id: 48,
-    nome: "Sebastian Sappy",
-    email: "ssappy1b@salon.com",
-    telefone: "7414883878",
-    codigoDaUnidadeConsumidora: "9131907229",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 5326.2,
-  },
-  {
-    id: 49,
-    nome: "Almeta Stodhart",
-    email: "astodhart1c@oakley.com",
-    telefone: "7014813339",
-    codigoDaUnidadeConsumidora: "6483938179",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 3133.35,
-  },
-  {
-    id: 50,
-    nome: "Berton Simmings",
-    email: "bsimmings1d@google.com.br",
-    telefone: "4041765087",
-    codigoDaUnidadeConsumidora: "2789697450",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 5852.11,
-  },
-  {
-    id: 51,
-    nome: "Roxanna Lauxmann",
-    email: "rlauxmann1e@psu.edu",
-    telefone: "3263976711",
-    codigoDaUnidadeConsumidora: "7193193678",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 3763.28,
-  },
-  {
-    id: 52,
-    nome: "Bealle d' Elboux",
-    email: "bd1f@pen.io",
-    telefone: "3731897061",
-    codigoDaUnidadeConsumidora: "7482665137",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 5309.27,
-  },
-  {
-    id: 53,
-    nome: "Kim Bootell",
-    email: "kbootell1g@ftc.gov",
-    telefone: "8969759945",
-    codigoDaUnidadeConsumidora: "8157427126",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 5975.53,
-  },
-  {
-    id: 54,
-    nome: "Kessiah Heindrich",
-    email: "kheindrich1h@creativecommons.org",
-    telefone: "7159039272",
-    codigoDaUnidadeConsumidora: "4458754750",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 8420.46,
-  },
-  {
-    id: 55,
-    nome: "Upton MacAne",
-    email: "umacane1i@hhs.gov",
-    telefone: "9543342807",
-    codigoDaUnidadeConsumidora: "1445204819",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 5656.82,
-  },
-  {
-    id: 56,
-    nome: "Alvy Radford",
-    email: "aradford1j@prlog.org",
-    telefone: "9272644087",
-    codigoDaUnidadeConsumidora: "1247034925",
-    enquadramento: "B1",
-    modeloFasico: "monofasico",
-    valor: 3438.74,
-  },
-  {
-    id: 57,
-    nome: "Juliette Borham",
-    email: "jborham1k@cnet.com",
-    telefone: "7194934961",
-    codigoDaUnidadeConsumidora: "5465773685",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 7667.15,
-  },
-  {
-    id: 58,
-    nome: "Valentina Eastlake",
-    email: "veastlake1l@youtu.be",
-    telefone: "5487945214",
-    codigoDaUnidadeConsumidora: "8523164766",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 9666.89,
-  },
-  {
-    id: 59,
-    nome: "Arabela Smallpeice",
-    email: "asmallpeice1m@newsvine.com",
-    telefone: "6159199740",
-    codigoDaUnidadeConsumidora: "2876169983",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 4801.83,
-  },
-  {
-    id: 60,
-    nome: "Trula Buzzing",
-    email: "tbuzzing1n@meetup.com",
-    telefone: "3052769796",
-    codigoDaUnidadeConsumidora: "0747089957",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 4667.55,
-  },
-  {
-    id: 61,
-    nome: "Tyler Canadas",
-    email: "tcanadas1o@51.la",
-    telefone: "8292136196",
-    codigoDaUnidadeConsumidora: "2238848568",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 8526.87,
-  },
-  {
-    id: 62,
-    nome: "Karlik Twigley",
-    email: "ktwigley1p@smh.com.au",
-    telefone: "4443257870",
-    codigoDaUnidadeConsumidora: "8773442305",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 6580.15,
-  },
-  {
-    id: 63,
-    nome: "Wyn Iashvili",
-    email: "wiashvili1q@omniture.com",
-    telefone: "1683965106",
-    codigoDaUnidadeConsumidora: "3449865552",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 6060.79,
-  },
-  {
-    id: 64,
-    nome: "Marieann Kuhn",
-    email: "mkuhn1r@stanford.edu",
-    telefone: "6651799916",
-    codigoDaUnidadeConsumidora: "4472239973",
-    enquadramento: "B1",
-    modeloFasico: "monofasico",
-    valor: 3677.86,
-  },
-  {
-    id: 65,
-    nome: "Evin Kettlesing",
-    email: "ekettlesing1s@google.com.hk",
-    telefone: "5394470368",
-    codigoDaUnidadeConsumidora: "5223046919",
-    enquadramento: "AX",
-    modeloFasico: "monofasico",
-    valor: 4152.85,
-  },
-  {
-    id: 66,
-    nome: "Dom Surridge",
-    email: "dsurridge1t@zdnet.com",
-    telefone: "2513087834",
-    codigoDaUnidadeConsumidora: "6676890062",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 9285.27,
-  },
-  {
-    id: 67,
-    nome: "Micky Gonet",
-    email: "mgonet1u@ted.com",
-    telefone: "5982131821",
-    codigoDaUnidadeConsumidora: "4362488529",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 6004.52,
-  },
-  {
-    id: 68,
-    nome: "Talia Biaggioni",
-    email: "tbiaggioni1v@last.fm",
-    telefone: "1546358777",
-    codigoDaUnidadeConsumidora: "2663497665",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 4235.48,
-  },
-  {
-    id: 69,
-    nome: "Raychel Sugge",
-    email: "rsugge1w@tmall.com",
-    telefone: "7492195344",
-    codigoDaUnidadeConsumidora: "1459965922",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 6988.12,
-  },
-  {
-    id: 70,
-    nome: "Dora Burchell",
-    email: "dburchell1x@dyndns.org",
-    telefone: "6792960021",
-    codigoDaUnidadeConsumidora: "9899582654",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 6012.16,
-  },
-  {
-    id: 71,
-    nome: "Zacharia Maldin",
-    email: "zmaldin1y@ed.gov",
-    telefone: "5959499345",
-    codigoDaUnidadeConsumidora: "1253523711",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 8626.95,
-  },
-  {
-    id: 72,
-    nome: "Stanley Roback",
-    email: "sroback1z@stumbleupon.com",
-    telefone: "3534717076",
-    codigoDaUnidadeConsumidora: "8754942268",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 7712.21,
-  },
-  {
-    id: 73,
-    nome: "Melloney Layton",
-    email: "mlayton20@ask.com",
-    telefone: "4971257951",
-    codigoDaUnidadeConsumidora: "6496042039",
-    enquadramento: "AX",
-    modeloFasico: "monofasico",
-    valor: 5050.92,
-  },
-  {
-    id: 74,
-    nome: "Phaedra Brouwer",
-    email: "pbrouwer21@jalbum.net",
-    telefone: "1072396345",
-    codigoDaUnidadeConsumidora: "7931891643",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 8375.86,
-  },
-  {
-    id: 75,
-    nome: "Debera Cunney",
-    email: "dcunney22@disqus.com",
-    telefone: "3193180793",
-    codigoDaUnidadeConsumidora: "6825877097",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 8831.23,
-  },
-  {
-    id: 76,
-    nome: "Gage McMurraya",
-    email: "gmcmurraya23@amazon.co.uk",
-    telefone: "2179519762",
-    codigoDaUnidadeConsumidora: "6030556959",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 6100.66,
-  },
-  {
-    id: 77,
-    nome: "Andy Sutliff",
-    email: "asutliff24@slate.com",
-    telefone: "6569266480",
-    codigoDaUnidadeConsumidora: "2307076549",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 7459.19,
-  },
-  {
-    id: 78,
-    nome: "Valencia Canape",
-    email: "vcanape25@skype.com",
-    telefone: "8742845569",
-    codigoDaUnidadeConsumidora: "6175377257",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 7149.1,
-  },
-  {
-    id: 79,
-    nome: "Earlie Benson",
-    email: "ebenson26@ucoz.ru",
-    telefone: "6827162831",
-    codigoDaUnidadeConsumidora: "6922718836",
-    enquadramento: "AX",
-    modeloFasico: "bifasico",
-    valor: 4023.92,
-  },
-  {
-    id: 80,
-    nome: "Berton McMurty",
-    email: "bmcmurty27@vinaora.com",
-    telefone: "4018715162",
-    codigoDaUnidadeConsumidora: "9932470120",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 9655.34,
-  },
-  {
-    id: 81,
-    nome: "Jacquie Fedynski",
-    email: "jfedynski28@dell.com",
-    telefone: "4129639806",
-    codigoDaUnidadeConsumidora: "4805811080",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 3553.24,
-  },
-  {
-    id: 82,
-    nome: "Lynna Ovendale",
-    email: "lovendale29@noaa.gov",
-    telefone: "7273947099",
-    codigoDaUnidadeConsumidora: "4585323880",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 5240.03,
-  },
-  {
-    id: 83,
-    nome: "Montgomery Bondley",
-    email: "mbondley2a@about.me",
-    telefone: "2489208303",
-    codigoDaUnidadeConsumidora: "3162209331",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 5755.68,
-  },
-  {
-    id: 84,
-    nome: "Page Fesby",
-    email: "pfesby2b@ustream.tv",
-    telefone: "6438772569",
-    codigoDaUnidadeConsumidora: "5468780356",
-    enquadramento: "AX",
-    modeloFasico: "bifasico",
-    valor: 6873.28,
-  },
-  {
-    id: 85,
-    nome: "Barnaby Secombe",
-    email: "bsecombe2c@meetup.com",
-    telefone: "2418922040",
-    codigoDaUnidadeConsumidora: "8427526261",
-    enquadramento: "AX",
-    modeloFasico: "bifasico",
-    valor: 6034.71,
-  },
-  {
-    id: 86,
-    nome: "Mureil Maciocia",
-    email: "mmaciocia2d@blogs.com",
-    telefone: "4847607788",
-    codigoDaUnidadeConsumidora: "0308738713",
-    enquadramento: "B2",
-    modeloFasico: "monofasico",
-    valor: 9629.09,
-  },
-  {
-    id: 87,
-    nome: "Ange Trittam",
-    email: "atrittam2e@seesaa.net",
-    telefone: "3637789092",
-    codigoDaUnidadeConsumidora: "1620714418",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 9507.85,
-  },
-  {
-    id: 88,
-    nome: "Kania Houndsom",
-    email: "khoundsom2f@nhs.uk",
-    telefone: "7124190409",
-    codigoDaUnidadeConsumidora: "3373939720",
-    enquadramento: "B1",
-    modeloFasico: "bifasico",
-    valor: 6964.72,
-  },
-  {
-    id: 89,
-    nome: "Eddy Rocca",
-    email: "erocca2g@cornell.edu",
-    telefone: "6291995441",
-    codigoDaUnidadeConsumidora: "2574468651",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 3602.48,
-  },
-  {
-    id: 90,
-    nome: "Christel Guillerman",
-    email: "cguillerman2h@mit.edu",
-    telefone: "5932585504",
-    codigoDaUnidadeConsumidora: "4980628605",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 8854.39,
-  },
-  {
-    id: 91,
-    nome: "Eddie Fountian",
-    email: "efountian2i@vinaora.com",
-    telefone: "7081919006",
-    codigoDaUnidadeConsumidora: "4493025417",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 3187.93,
-  },
-  {
-    id: 92,
-    nome: "Cassi Girvan",
-    email: "cgirvan2j@boston.com",
-    telefone: "2325305785",
-    codigoDaUnidadeConsumidora: "0832211508",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 3467.0,
-  },
-  {
-    id: 93,
-    nome: "Sam Jandl",
-    email: "sjandl2k@yale.edu",
-    telefone: "2152968290",
-    codigoDaUnidadeConsumidora: "5955905944",
-    enquadramento: "AX",
-    modeloFasico: "trifasico",
-    valor: 5252.71,
-  },
-  {
-    id: 94,
-    nome: "Ara Grichukhin",
-    email: "agrichukhin2l@vistaprint.com",
-    telefone: "7789310069",
-    codigoDaUnidadeConsumidora: "3650524201",
-    enquadramento: "B2",
-    modeloFasico: "trifasico",
-    valor: 7735.45,
-  },
-  {
-    id: 95,
-    nome: "Thain Tubridy",
-    email: "ttubridy2m@cbslocal.com",
-    telefone: "3731085366",
-    codigoDaUnidadeConsumidora: "4139000457",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 8142.23,
-  },
-  {
-    id: 96,
-    nome: "Eliot Burlingame",
-    email: "eburlingame2n@yale.edu",
-    telefone: "7105428173",
-    codigoDaUnidadeConsumidora: "6779766437",
-    enquadramento: "B3",
-    modeloFasico: "bifasico",
-    valor: 9389.73,
-  },
-  {
-    id: 97,
-    nome: "Jeanne Woliter",
-    email: "jwoliter2o@narod.ru",
-    telefone: "2889074583",
-    codigoDaUnidadeConsumidora: "9830718522",
-    enquadramento: "B2",
-    modeloFasico: "bifasico",
-    valor: 6816.16,
-  },
-  {
-    id: 98,
-    nome: "Giana Seivwright",
-    email: "gseivwright2p@seattletimes.com",
-    telefone: "1841782726",
-    codigoDaUnidadeConsumidora: "7745845742",
-    enquadramento: "B3",
-    modeloFasico: "trifasico",
-    valor: 9655.23,
-  },
-  {
-    id: 99,
-    nome: "Jorry Maffy",
-    email: "jmaffy2q@ning.com",
-    telefone: "9334771556",
-    codigoDaUnidadeConsumidora: "8605387076",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 6698.2,
-  },
-  {
-    id: 100,
-    nome: "Ogdon Whitwham",
-    email: "owhitwham2r@linkedin.com",
-    telefone: "1148543026",
-    codigoDaUnidadeConsumidora: "8717051940",
-    enquadramento: "B3",
-    modeloFasico: "monofasico",
-    valor: 5693.12,
-  },
-];
